requirements.txt
bidict==0.22.1
bingX-connector==0.0.7
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
dnspython==2.3.0
eventlet==0.33.3
Flask==2.3.2
aioflask
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
Flask-SocketIO==5.3.4
greenlet==2.0.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
netifaces==0.11.0
pymongo==4.4.1
python-dotenv==1.0.0
python-engineio==4.5.1
python-socketio==5.8.0
requests==2.31.0
six==1.16.0
urllib3==1.26.15
waitress==2.1.2
Werkzeug==2.3.6
zipp==3.16.2
ejtraderCT==1.1.13